refactor into a single binary
